fix: restructure providers, separate request/response conversion (#15)
