+++ src/providers/anthropic/mod.rs
@@ -1,0 +1,4 @@
+mod models;
+mod provider;
+
+pub use provider::AnthropicProvider;
