+++ src/providers/azure/mod.rs
@@ -1,0 +1,3 @@
+mod provider;
+
+pub use provider::AzureProvider;
